+++ ls_Whistles/locales/enUS.lua
@@ -5,7 +5,7 @@ local C, D, L = addon.C, addon.D, addon.L
 local _G = getfenv(0)
 
 -- Mine
-L["LS_WHISTLES"] = ("LS: |c%sWhistles|r"):format(D.global.colors.addon:GetHex())
+L["ADDON_NAME"] = ("LS: |c%sWhistles|r"):format(D.global.colors.addon:GetHex())
 L["CURSEFORGE"] = "CurseForge"
 L["DISCORD"] = "Discord"
 L["GITHUB"] = "GitHub"
@@ -14,10 +14,12 @@ L["WOWINTERFACE"] = "WoWInterface"
 L["INFO"] = D.global.colors.addon:WrapTextInColorCode(_G.INFO)
 L["AC_TOOLTIP"] = ("|c%1$sClick:|r %2$s\n|c%1$sShift Click:|r %3$s"):format(D.global.colors.addon:GetHex(), _G.GAMEMENU_OPTIONS, _G.ADVANCED_OPTIONS)
 
+L["IMPORT"] = _G.HUD_EDIT_MODE_IMPORT_LAYOUT
+
 -- Require translation
-L["ACTIONBARS_DESC"] = "Adds bar fading and button text options for each individual action bar in Edit Mode."
+L["ACTIONBARS_FOOTER"] = "Also adds bar fading and button text options for each individual action bar in Edit Mode."
 L["AJ_TAB_DESC"] = "Makes the tab more information dense."
-L["BACKPACK_DESC"] = "Adds tracked currencies, gold, and the current token price to the tooltip."
+L["BACKPACK_DESC"] = "Adds tracked currencies, gold, and the current token price to the tooltip. Also adds fading options in Edit Mode."
 L["BONUS_XP_TOOLTIP"] = "Bonus XP: |cffffffff%s|r"
 L["BORDER"] = "Border"
 L["CAST_VFX"] = "Casting Visual Effects"
@@ -33,12 +35,14 @@ L["DESATURATION"] = "Desaturation"
 L["DOWNLOADS"] = "Downloads"
 L["EJ_BUTTON_DESC"] = "Adds raid and dungeon lockout information."
 L["EQUIPPED_HIGHLIGHT"] = "Usable Equipment Highlight"
+L["EXPORT"] = "Export"
 L["FADING"] = "Fading"
 L["FADING_COMBAT_DESC"] = "Fade in on entering combat."
 L["FADING_TARGET_DESC"] = "Fade in on acquiring target."
 L["FONT"] = "Font"
 L["GRYPHON"] = "Gryphon"
 L["INSPECT_FRAME"] = "Inspect Frame"
+L["ITEM_COUNT"] = "Item Count"
 L["JOURNEYS_TOOLTIP_WATCH_JOURNEY"] = "<Shift Click to Show as Experience Bar>"
 L["LATENCY"] = "Latency"
 L["LATENCY_HOME"] = "Home"
@@ -52,6 +56,7 @@ L["LOOT_DESC"] = "Makes the loot window significantly more compact and adds a bu
 L["MAIL_DESC"] = "Adds a button to remove all empty messages."
 L["MAINMENU_BUTTON_DESC"] = "Add addon performance information."
 L["MAINMENU_BUTTON_HOLD_TOOLTIP"] = "|cffffffffHold Shift|r to show performance information."
+L["MICROMENU_FOOTER"] = "Also adds fading options in Edit Mode."
 L["MIN_ALPHA"] = "Min Alpha"
 L["MISSING_ENCHANTS"] = "Missing Enhancements"
 L["NUMBERS"] = "Numbers"
@@ -66,4 +71,5 @@ L["SHIFT_CLICK_TO_SHOW_AS_XP"] = "|cffffffffShift click|r to show as experience 
 L["SHORT_HOTKEY"] = "Short Key Bindings"
 L["SUPPORT_FEEDBACK"] = "Support & Feedback"
 L["TEXTURE"] = "Texture"
+L["TOOLTIP_IDS"] = "Spell and Item IDs"
 L["WYVERN"] = "Wyvern"